Refactor run package
