+++ cli/command/run/xcuitest.go
@@ -1,0 +1,88 @@
+package run
+
+import (
+	"errors"
+	"fmt"
+	"github.com/rs/zerolog/log"
+	"github.com/saucelabs/saucectl/internal/appstore"
+	"github.com/saucelabs/saucectl/internal/rdc"
+	"github.com/saucelabs/saucectl/internal/region"
+	"github.com/saucelabs/saucectl/internal/resto"
+	"github.com/saucelabs/saucectl/internal/saucecloud"
+	"github.com/saucelabs/saucectl/internal/testcomposer"
+	"github.com/saucelabs/saucectl/internal/xcuitest"
+	"github.com/spf13/cobra"
+)
+
+func runXcuitest(cmd *cobra.Command, tc testcomposer.Client, rs resto.Client, rc rdc.Client, as *appstore.AppStore) (int, error) {
+	p, err := xcuitest.FromFile(gFlags.cfgFilePath)
+	if err != nil {
+		return 1, err
+	}
+	p.Sauce.Metadata.ExpandEnv()
+	applyDefaultValues(&p.Sauce)
+	overrideCliParameters(cmd, &p.Sauce, &p.Artifacts)
+
+	regio := region.FromString(p.Sauce.Region)
+	if regio == region.None {
+		log.Error().Str("region", gFlags.regionFlag).Msg("Unable to determine sauce region.")
+		return 1, errors.New("no sauce region set")
+	}
+
+	xcuitest.SetDeviceDefaultValues(&p)
+	err = xcuitest.Validate(p)
+	if err != nil {
+		return 1, err
+	}
+
+	if cmd.Flags().Lookup("suite").Changed {
+		if err := filterXcuitestSuite(&p); err != nil {
+			return 1, err
+		}
+	}
+
+	tc.URL = regio.APIBaseURL()
+	rs.URL = regio.APIBaseURL()
+	as.URL = regio.APIBaseURL()
+	rc.URL = regio.APIBaseURL()
+
+	rs.ArtifactConfig = p.Artifacts.Download
+	rc.ArtifactConfig = p.Artifacts.Download
+
+	return runXcuitestInCloud(p, regio, tc, rs, rc, as)
+}
+
+func runXcuitestInCloud(p xcuitest.Project, regio region.Region, tc testcomposer.Client, rs resto.Client, rc rdc.Client, as *appstore.AppStore) (int, error) {
+	log.Info().Msg("Running XCUITest in Sauce Labs")
+	printTestEnv("sauce")
+
+	r := saucecloud.XcuitestRunner{
+		Project: p,
+		CloudRunner: saucecloud.CloudRunner{
+			ProjectUploader:       as,
+			JobStarter:            &tc,
+			JobReader:             &rs,
+			RDCJobReader:          &rc,
+			JobStopper:            &rs,
+			JobWriter:             &tc,
+			CCYReader:             &rs,
+			TunnelService:         &rs,
+			Region:                regio,
+			ShowConsoleLog:        false,
+			ArtifactDownloader:    &rs,
+			RDCArtifactDownloader: &rc,
+			DryRun:                gFlags.dryRun,
+		},
+	}
+	return r.RunProject()
+}
+
+func filterXcuitestSuite(c *xcuitest.Project) error {
+	for _, s := range c.Suites {
+		if s.Name == gFlags.suiteName {
+			c.Suites = []xcuitest.Suite{s}
+			return nil
+		}
+	}
+	return fmt.Errorf("suite name '%s' is invalid", gFlags.suiteName)
+}
